backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "smartfarm-backend"
version = "1.0.0"
description = "AI-powered agriculture assistant backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.4",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "numpy>=1.26.4",
    "scikit-learn>=1.4.1.post1",
    "python-dotenv>=1.0.1",
]
